Move json.go from utils/ to excellent/types/
