+++ excellent/types/conversions_test.go
@@ -126,14 +126,14 @@ func TestToBool(t *testing.T) {
 		{"lO.5", true, false},
 		{"", false, false},
 		{testResolver, true, false},
-		{utils.JSONFragment([]byte(`false`)), false, false},
-		{utils.JSONFragment([]byte(`true`)), true, false},
-		{utils.JSONFragment([]byte(`[]`)), false, false},
-		{utils.JSONFragment([]byte(`15.5`)), true, false},
-		{utils.JSONFragment([]byte(`0`)), false, false},
-		{utils.JSONFragment([]byte(`[5]`)), true, false},
-		{utils.JSONFragment([]byte("{\n}")), false, false},
-		{utils.JSONFragment([]byte(`{"one": "two"}`)), true, false},
+		{types.JSONFragment([]byte(`false`)), false, false},
+		{types.JSONFragment([]byte(`true`)), true, false},
+		{types.JSONFragment([]byte(`[]`)), false, false},
+		{types.JSONFragment([]byte(`15.5`)), true, false},
+		{types.JSONFragment([]byte(`0`)), false, false},
+		{types.JSONFragment([]byte(`[5]`)), true, false},
+		{types.JSONFragment([]byte("{\n}")), false, false},
+		{types.JSONFragment([]byte(`{"one": "two"}`)), true, false},
 		{struct{}{}, false, true},
 	}
