+++ flows/triggers/base.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/nyaruka/goflow/excellent/types"
 	"github.com/nyaruka/goflow/flows"
 	"github.com/nyaruka/goflow/utils"
 )
@@ -12,14 +13,14 @@ type baseTrigger struct {
 	environment utils.Environment
 	flow        flows.Flow
 	contact     *flows.Contact
-	params      utils.JSONFragment
+	params      types.JSONFragment
 	triggeredOn time.Time
 }
 
 func (t *baseTrigger) Environment() utils.Environment { return t.environment }
 func (t *baseTrigger) Flow() flows.Flow               { return t.flow }
 func (t *baseTrigger) Contact() *flows.Contact        { return t.contact }
-func (t *baseTrigger) Params() utils.JSONFragment     { return t.params }
+func (t *baseTrigger) Params() types.JSONFragment     { return t.params }
 func (t *baseTrigger) TriggeredOn() time.Time         { return t.triggeredOn }
 
 // Resolve resolves the given key when this trigger is referenced in an expression
