+++ flows/triggers/manual.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"time"
 
+	"github.com/nyaruka/goflow/excellent/types"
 	"github.com/nyaruka/goflow/flows"
 	"github.com/nyaruka/goflow/utils"
 )
@@ -29,7 +30,7 @@ type ManualTrigger struct {
 }
 
 // NewManualTrigger creates a new manual trigger
-func NewManualTrigger(env utils.Environment, contact *flows.Contact, flow flows.Flow, params utils.JSONFragment, triggeredOn time.Time) flows.Trigger {
+func NewManualTrigger(env utils.Environment, contact *flows.Contact, flow flows.Flow, params types.JSONFragment, triggeredOn time.Time) flows.Trigger {
 	return &ManualTrigger{baseTrigger{environment: env, contact: contact, flow: flow, triggeredOn: triggeredOn}}
 }
